refactor(auto_search): extract release_parse.rs (regex-heavy parsers)
Fourth slice. Extracted the release-filename token parsing cluster:

  static RE_EPISODE_PATTERNS, RE_RANGE, RE_NTH_SEASON, RE_SEASON_N,
         RE_PART_COUR, RE_SXXEXX, RE_STANDALONE_S, RE_RELEASE_SEASON_N,
         RE_RELEASE_NTH_SEASON, RE_RELEASE_PART_COUR, RE_RELEASE_ROMAN,
         RE_PART_N, RE_ROMAN_PART

  fn is_noise_number
  pub fn parse_release_numbers
  pub fn infer_season_from_detail
  pub(super) fn infer_season_from_title
  pub fn parse_release_season
  fn roman_to_i32 / roman_to_int
  pub fn extract_part_number
  pub fn pick_wanted_file_indices (+ helpers pick_by_part_number / pick_by_subtitle_include)
  within_expected_episode_count / pub(super) within_episode_slack
  pub fn extract_season_subtitle / pub(super) trailing_subtitle_of / is_generic_season_subtitle
  pub fn has_selective_discriminator
  pub(super) fn normalize_subtitle
  pub(crate) fn is_media_filename
  pub(super) fn season_mismatch

  + 26 #[test] fns covering all of the above, plus a `detail_with_titles`
    helper duplicated here (also kept in mod.rs since pack_detection tests
    use it too — will consolidate when pack_detection extracts).

Module wiring: release_parse.rs imports `super::collect_aliases`
(used by infer_season_from_detail). mod.rs re-exports the public
surface via `pub use release_parse::{…}` and `pub(crate) use
release_parse::is_media_filename;` (crate-local visibility preserved
from the original).

CI pass with 2 pub(super) bumps and 1 manual visibility fix applied:
  .infer_season_from_title → pub(super) (used by infer_season_from_detail)
  .trailing_subtitle_of     → pub(super) (used by pack-detection code in mod.rs)
  .within_episode_slack     → pub(super) (same)
  .normalize_subtitle       → pub(super) (same)

Also cleaned up:
  - orphan `#[derive(Debug, Clone)]` left over from a moved struct
  - orphan `#[test]` that was pointing at a helper fn post-extract
  - unused `use regex_lite::Regex;` in mod.rs after regexes moved

mod.rs: 5316 → 4300 lines. release_parse.rs: 1074 lines.
Tests still passing: 601.
